expanded some of the reused stuff into components
