+++ app/helpers/interface.ts
@@ -1,16 +1,16 @@
 import { Dispatch, SetStateAction } from "react";
-import { Blog, PreBlog, User } from "../types";
+import { BlogI, PreBlog, User } from "../types";
 
 export interface editBlogI { 
   setPreBlog: Dispatch<SetStateAction<PreBlog>>
-  blog: Blog
+  blog: BlogI
 }
 export interface updateBlogI { 
   setIsOpen: Dispatch<SetStateAction<boolean>>
   preBlog: PreBlog
   setPreBlog: Dispatch<SetStateAction<PreBlog>>
-  originalBlog: Blog | undefined
-  setOriginalBlog: Dispatch<SetStateAction<Blog | undefined>>
+  originalBlog: BlogI | undefined
+  setOriginalBlog: Dispatch<SetStateAction<BlogI | undefined>>
   setCreatingBlog: Dispatch<SetStateAction<boolean>>
   user: User
 }
