+++ app/page.tsx
@@ -1,5 +1,5 @@
 'use client'
-import React, { Dispatch, SetStateAction, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import Image from 'next/image'
 import "react-markdown-editor-lite/lib/index.css";
 import dynamic from 'next/dynamic';
@@ -7,14 +7,11 @@ import ReactMarkdown from 'react-markdown'
 import remarkGfm from 'remark-gfm'
 import 'github-markdown-css/github-markdown-light.css'
 import Axios from 'axios';
-import {API, DATE_TYPE, ENV, S3_URL} from './constants'
+import {API, S3_URL} from './constants'
 import {
-  addOrRemoveTag,
-  dataURLtoFile, editBlog,
-  getDateAge, isEmpty,
+  addOrRemoveTag, editBlog, getDateAge, isEmpty,
   sortAndReduce, sortSagaFilters,
-  sortTagFilters, toggleSagaFilter,
-  toggleTagFilter, handleImageUpload
+  sortTagFilters, handleImageUpload
 } from './helpers/helpers';
 import { Amplify, Auth } from 'aws-amplify';
 import { userPool } from './constants';
@@ -22,18 +19,13 @@ import { userPool } from './constants';
 import ArrowLeft from './assets/ArrowLeft';
 import ArrowRight from './assets/ArrowRight';
 import Plus from './assets/Plus';
-import Remove from './assets/Remove';
-import { Saga, User, PreBlog, Blog, FilterBlog } from './types';
+import { Saga, User, PreBlog, BlogI, FilterBlog } from './types';
 import ArrowDown from './assets/ArrowDown';
 import InfiniteScroll from 'react-infinite-scroll-component';
 import Modal from 'react-modal';
-import MoreDots from './assets/MoreDots';
-import { api, helpers } from './helpers';
-import { deleteOrHideBlog, updateBlog } from './helpers/api';
-import Edit from './assets/Edit';
-import Eye from './assets/Eye';
-import Bin from './assets/Bin';
-import EyeOff from './assets/EyeOff';
+
+import { Blog, Bubble, SagaFilter, TagFilter } from '@/components';
+import ModalComponent from '@/components/Modal';
 
 Amplify.configure({
   Auth: {
@@ -53,24 +45,22 @@ const PAGE_SIZE = 5
 let tagsLength = 0;
 let sagasLength = 0;
 let blogsLength = 0;
-let storageArray: Blog[] = []
+let storageArray: BlogI[] = []
 export default function Home() {
   const [currentUser, setCurrentUser] = useState<User>({location: '', bio: '', createdAt: '', id: '', name: '', sagas: [], tags: [], type: '', profileImage: '', draft: ''})
   const [pageAuthor, setPageAuthor] = useState<User>({location: '', bio: '', createdAt: '', id: '', name: '', sagas: [], tags: [], type: '', profileImage: ''})
   const [preBlog, setPreBlog] = useState<PreBlog>({title: '', body: '', userId: currentUser.id, author: currentUser.name, tags: [], saga: ''});
-  const [originalBlog, setOriginalBlog] = useState<Blog | undefined>(undefined);
-  const [blogs, setBlogs] = useState<Blog[]>([]);
+  const [originalBlog, setOriginalBlog] = useState<BlogI | undefined>(undefined);
+  const [blogs, setBlogs] = useState<BlogI[]>([]);
   const [authenticated, setAuthenticated] = useState(false)
   const [loaded, setLoaded] = useState(false)
   const [creatingBlog, setCreatingBlog] = useState(false)
   const [isEditing, setIsEditing] = useState(false)
   const [isOpen, setIsOpen] = useState(false)
-  const [actionModalVisible, setActionModalVisible] = useState(false)
 
   const [loggingIn, setLoggingIn] = useState(false)
   const [email, setEmail] = useState('')
   const [password, setPassword] = useState('')
-  const [tag, setTag] = useState('')
 
   const [pageSagas, setPageSagas] = useState<Saga[]>([]);
   const [pageTags, setPageTags] = useState<string[]>([]);
@@ -146,35 +136,7 @@ export default function Home() {
     }
   }
 
-  const createBlog = () => {
-    if (isEditing) {
-      updateBlog({ setIsOpen, setPreBlog, originalBlog, preBlog, setOriginalBlog, setCreatingBlog, user: currentUser });
-      setIsEditing(false)
-    } else {
-      setIsOpen(false);
 
-      let newSagas = currentUser.sagas || [];
-      if (!isEmpty(preBlog.saga)) {
-        const index = newSagas.findIndex(t => t.saga.toLowerCase() === preBlog.saga.toLowerCase());
-        if (index >= 0) {
-          newSagas[index].updated = '';
-        } else if (index === -1) {
-          newSagas.push({saga: preBlog.saga, updated: ''})
-        }
-      }
-
-      const newTags = preBlog.tags.filter((str: string) => !isEmpty(str));
-      let combinedTags = Array.from(new Set(currentUser.tags.concat(newTags)));
-      if (combinedTags === currentUser.tags) {
-        combinedTags = []
-      }
-
-      Axios.post(`${API}/createBlog`, { ...preBlog, userTags: combinedTags, userSagas: newSagas, createdAt: currentUser.createdAt }).then(res => {
-        setPreBlog({ title: '', body: '', userId: currentUser.id, author: currentUser.name, tags: [], saga: '' });
-        setCreatingBlog(false);
-      })
-    }
-  }
 
   const getBlogsFiltered = () => {
     if (filterSaga === '' && filterTags.length === 0) {
@@ -238,8 +200,8 @@ export default function Home() {
   }
 
   const blogItem = blogs?.map((item) => (
-    <BlogItem key={item.id} blog={item} owned={item.userId === currentUser.id} setPreBlog={setPreBlog}
-      preBlog={preBlog} setBlogs={setBlogs} blogs={blogs} />
+    <Blog key={item.id} blog={item} owned={item.userId === currentUser.id} setPreBlog={setPreBlog} preBlog={preBlog} setBlogs={setBlogs}
+      blogs={blogs} pageAuthor={pageAuthor} currentUser={currentUser} setOriginalBlog={setOriginalBlog} setCreatingBlog={setCreatingBlog}/>
   ))
 
   if (!loaded) {
@@ -271,51 +233,9 @@ export default function Home() {
                 <button onClick={() => setIsOpen(true)} disabled={preBlog.body === ''} className={`${preBlog.body === '' ? 'bg-sky-200' : 'bg-sky-300'} px-8 py-2 rounded-full text-neutral-50 font-bold`}>confirm</button>
               </div>
             </div>
-            <Modal isOpen={isOpen} onRequestClose={() => setIsOpen(false)} ariaHideApp={false}
-              className='border-0 left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 w-1/3 fixed'>
-              <div className='flex-col flex w-full bg-white p-5 shadow-md rounded-2xl'>
-                <div className='w-full py-1 flex'>
-                  <span className='font-bold text-sky-300 mr-3 w-10'>Title</span>
-                  <input className='border-b-neutral-200 border-b-2 focus:border-sky-300 focus:outline-none' value={preBlog.title}
-                    type="text" placeholder='title' onChange={(e) => setPreBlog(prev => ({ ...prev, title: e.target.value }))} />
-                </div>
-                <div className='w-full flex my-6'>
-                  <span className='font-bold text-sky-300 mr-3 w-10'>Saga</span>
-                  <input disabled={isEditing} className='border-b-neutral-200 border-b-2 focus:border-sky-300 focus:outline-none' value={preBlog.saga}
-                    type="text" placeholder='saga' onChange={(e) => setPreBlog(prev => ({ ...prev, saga: e.target.value }))} />
-                </div>
-                <div className='w-full flex'>
-                  <span className='font-bold text-sky-300 mr-3 w-10'>Tags</span>
-                  <div className='flex flex-row items-center'>
-                    <input disabled={isEditing} className='border-b-neutral-200 border-b-2 focus:border-sky-300 focus:outline-none mr-3' type="text" placeholder='tag' onChange={(e) => setTag(e.target.value)} />
-                    <Plus onClick={() => { !isEditing ? setPreBlog(prev => ({ ...prev, tags: [...prev.tags, tag] })) : null; setTag('') }} width={20}/> 
-                  </div>
-                </div>
-                <span className={`${preBlog.tags.length>0 ? '' : 'hidden'} text-sm text-neutral-600 mt-5 mb-1`}>current tags</span>
-                <div className={`${preBlog.tags.length > 0 ? '' : 'hidden'} flex flex-row w-full flex-wrap mb-3`}>
-                  {preBlog.tags.map((tag) => (
-                    <Bubble key={tag} name={tag} preBlog={preBlog} type='tag' setPreBlog={setPreBlog}/>
-                  ))}
-                </div>
-                <span className='text-sm text-neutral-600 mt-5 mb-1'>previous tags</span>
-                <div className='flex flex-row w-full flex-wrap mb-3'>
-                  {pageAuthor.tags.map((tag) => (
-                    <div key={tag} onClick={() => addOrRemoveTag({ tag: tag, preBlog: preBlog, setPreBlog: setPreBlog })}
-                      className={`flex-row flex items-center mr-3 mb-2 px-3 py-1 border-1 rounded-full border-sky-300 cursor-pointer select-none ${preBlog.tags.includes(tag) ? 'hidden' : ''}`}>
-                      <span className='mr-1 text-neutral-500'>{tag}</span>
-                      <span><Plus width={15} /></span>
-                    </div>
-                  ))}
-                </div>
-                <div className='flex-row flex justify-between'>
-                  <button onClick={() => setIsOpen(false) }
-                    className='bg-neutral-200 px-8 py-2 rounded-full text-neutral-400 font-bold'>back</button>
-                  <button onClick={() => createBlog()} disabled={preBlog.title === '' || preBlog.saga === ''}
-                    className={`${preBlog.title === '' || preBlog.saga === '' ? 'bg-sky-200' : 'bg-sky-300'} px-8 py-2 rounded-full text-neutral-50 font-bold`}>{isEditing ? 'update' : 'create'}</button>
-                </div>
-              </div>
-            </Modal>
-
+            <ModalComponent setIsOpen={setIsOpen} isOpen={isOpen} preBlog={preBlog} setPreBlog={setPreBlog} blogs={blogs}
+              setBlogs={setBlogs} isEditing={isEditing} setIsEditing={setIsEditing} pageAuthor={pageAuthor} currentUser={currentUser} orginalBlog={originalBlog}
+              setOriginalBlog={setOriginalBlog} setCreatingBlog={setCreatingBlog} />
           </div>) : (null)}
           <InfiniteScroll
             dataLength={blogs?.length || 0}
@@ -435,167 +355,4 @@ export default function Home() {
       </div>
     )
   }
-}
-
-interface BlogProps {
-  blog: Blog
-  owned: boolean
-  preBlog: PreBlog
-  setPreBlog: Dispatch<SetStateAction<PreBlog>>
-  blogs: Blog[]
-  setBlogs: Dispatch<SetStateAction<Blog[]>>
-}
-
-const BlogItem: React.FC<BlogProps> = ({ blog, owned, setPreBlog, preBlog, blogs, setBlogs }) => {
-  const [eyeHover, setEyeHover] = useState(false)
-  const [isOpen, setIsOpen] = useState(false)
-  const [isEditing, setIsEditing] = useState(false)
-
-  const toggleVisibility = (state: boolean) => {
-    setEyeHover(state);
-    const tempBlogs = blogs;
-    tempBlogs[blogs.findIndex(b => b.id === blog.id)].visible = state;
-    setBlogs(tempBlogs);
-    deleteOrHideBlog(false, blog.visible, blog.id, blog.createdAt)
-  }
-  const deleteBlog = () => {
-    console.log("deleted!")
-    const tempBlogs = blogs.filter(b => b.id !== blog.id);
-    setBlogs(tempBlogs);
-    setIsOpen(false);
-    // deleteBlog(true, false, blog.id, blog.createdAt)
-  }
-
-  return (
-    <div className='flex-col flex mb-20'>
-      <div className='flex-row flex justify-between items-center'>
-      <div className='flex-col flex justify-between'>
-        <span className='text-xl font-semibold'>{blog.title}</span>
-        <div className='flex-row'>
-            <span title={blog.editedAt ? getDateAge(blog.editedAt, DATE_TYPE.EDIT) : ''}
-              className='text-sm mt-1'>{getDateAge(blog.createdAt, DATE_TYPE.BLOG)} {blog.editedAt ? '*' : ''}</span>
-        </div>
-        </div>
-        <div className='flex-row flex items-center'>
-          <Bubble name={blog.saga} type='saga' />
-          {owned ? (
-            <div className='flex-row flex pl-3'>
-              <Edit className='cursor-pointer' stroke='#3072B4' width={25}
-                onClick={() => { editBlog({ setPreBlog, blog }); setIsEditing(!isEditing) }} />
-              {blog.visible && !eyeHover || !blog.visible && eyeHover ?
-                (<Eye className='mx-3 cursor-pointer' stroke='#FF7A00' width={25} 
-                  onMouseEnter={() => setEyeHover(true)} onMouseLeave={() => setEyeHover(false)}
-                  onClick={() => toggleVisibility(true)} />) :
-                (<EyeOff className='mx-3 cursor-pointer' stroke='#FF7A00' width={25}
-                  onMouseEnter={() => setEyeHover(true)} onMouseLeave={() => setEyeHover(false)}
-                  onClick={() => toggleVisibility(false)} />)}
-              <Bin className='cursor-pointer' stroke='#dc2626' width={25}
-                onClick={() => setIsOpen(true)} />
-              {/* <MoreDots onClick={() => { editBlog({ setPreBlog, blog, setCreatingBlog, setOriginalBlog }); setIsEditing(true) }}
-                className='ml-3 cursor-pointer' width={20} /> */}
-            </div>
-          ) : (null)}
-        </div>
-      </div>
-      {isEditing ? (<div className={`${blog.tags.length > 0 ? 'mb-10' : ''}`}>
-            <div className='border-sky-300 border-2 rounded-2xl overflow-clip flex h-fit max-h-full my-5'>
-              <MdEditor
-                value={preBlog.body}
-                allowPasteImage
-                onImageUpload={handleImageUpload}
-                view={{ menu: true, html: false, md: true }}
-                canView={{ menu: true, html: true, both: false, fullScreen: false, hideMenu: false, md: true }}
-                placeholder='blog loblaw'
-                onChange={(text) => setPreBlog(prev => ({ ...prev, body: text.text }))}
-                plugins={['mode-toggle', 'link', 'block-code-inline', 'font-strikethrough', 'font-bold', 'font-italic', 'divider', 'block-code-block', 'block-quote', 'list-unordered', 'list-ordered', 'image', 'block-wrap']}
-                className='flex flex-grow rounded-2xl border-none h-fit max-h-full min-h-500 max-w-full'
-                renderHTML={text => <ReactMarkdown remarkPlugins={[remarkGfm]} linkTarget={'_blank'}>{text}</ReactMarkdown>}
-              />
-            </div>
-            <div className='flex-row flex justify-between mt-3'>
-              <button onClick={() => { setIsEditing(false)}} className='bg-neutral-200 px-8 py-2 rounded-full text-neutral-400 font-bold'>cancel</button>
-              <div>
-                {/* <button hidden={isEditing} onClick={() => saveDraft()} disabled={preBlog.body === currentUser.draft} className={`${preBlog.body !== currentUser.draft ? 'bg-sky-300' : 'bg-sky-200'} px-8 mr-3 py-2 rounded-full text-neutral-50 font-bold`}>save</button> */}
-                <button onClick={() => setIsOpen(true)} disabled={preBlog.body === ''} className={`${preBlog.body === '' ? 'bg-sky-200' : 'bg-sky-300'} px-8 py-2 rounded-full text-neutral-50 font-bold`}>confirm</button>
-              </div>
-            </div></div>) : (<ReactMarkdown className={`${blog.tags.length > 0 ? 'mb-5' : ''} markdown`} remarkPlugins={[remarkGfm]} linkTarget={'_blank'}>{blog.body}</ReactMarkdown>)}
-      
-      <div className='flex-row flex flex-wrap'>
-        {blog.tags.map((tag) => {
-          return <Bubble key={tag} name={tag} type='tag'/>
-        })}
-      </div>
-      <Modal isOpen={isOpen} onRequestClose={() => setIsOpen(false)} ariaHideApp={false}
-              className='border-0 left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 w-1/3 fixed focus-visible:outline-none'>
-              <div className='flex-col flex w-full bg-white p-8 shadow-md rounded-2xl'>
-          <p>{`Delete blog "${blog.title}" ? This cannot be undone`}</p>
-          <div className='flex flex-row items-center h-12 justify-between mt-3'>
-            <span className='cursor-pointer border-2 border-red-600 rounded-full w-28 p-2 text-center font-medium text-red-600'
-              onClick={() => deleteBlog()}>yes</span>
-            <span className='cursor-pointer border-2 border-sky-300 rounded-full w-28 p-2 text-center font-medium text-sky-300'
-              onClick={() => setIsOpen(false)}>no</span>
-          </div>
-
-              </div>
-            </Modal>
-    </div>
-  )
-}
-
-interface SagaProps {
-  name: string,
-  updated: string,
-  filterSaga: string
-  setFilterSaga: Dispatch<SetStateAction<string>>
-}
-const SagaFilter: React.FC<SagaProps> = ({ name, updated, filterSaga, setFilterSaga }) => {
-  return (
-    <div onClick={() => toggleSagaFilter({name, filterSaga, setFilterSaga})} className='flex-row flex w-full justify-between my-2 cursor-pointer'>
-      <span className={`font-medium ${filterSaga===name ? 'text-sky-300' : ''}`}>{name}</span>
-      <span className={`font-medium ${filterSaga===name ? 'text-sky-300' : ''}`}>{getDateAge(updated,DATE_TYPE.SAGA)}</span>
-    </div>
-  )
-}
-
-interface TagProps {
-  name: string,
-  filterTags: string[]
-  setFilterTags: Dispatch<SetStateAction<string[]>>
-}
-const TagFilter: React.FC<TagProps> = ({ name, filterTags, setFilterTags }) => {
-  return (
-    <div onClick={() => toggleTagFilter({ name, filterTags, setFilterTags })} className='flex-row flex w-full justify-between my-2 cursor-pointer'>
-      <span className={`font-medium ${filterTags.includes(name) ? 'text-sky-300' : ''}`}>{name}</span>
-      {filterTags.includes(name) ? <Remove width={15} /> : <Plus width={15} />}
-    </div>
-  )
-}
-
-interface Bubble {
-  name: string
-  type: string
-  preBlog?: PreBlog,
-  setPreBlog?: Dispatch<SetStateAction<PreBlog>>
-}
-const Bubble: React.FC<Bubble> = ({ name, type, preBlog, setPreBlog }) => {
-  if (preBlog && setPreBlog) {
-    return (
-      <div onClick={() => addOrRemoveTag({ tag: name, preBlog: preBlog, setPreBlog: setPreBlog })}
-        className={`flex-row flex items-center mr-3 mb-2 px-3 py-1 border-1 rounded-full border-sky-300 cursor-pointer select-none ${preBlog.tags.includes(name) ? 'bg-sky-50' : ''}`}>
-        <span className='mr-1 text-sky-300'>{name}</span>
-        <span>{preBlog.tags.includes(name) ? <Remove width={15} /> : <Plus width={15} />}</span>
-      </div>
-    )
-  } else if (type === 'tag') {
-    return (
-        <div className='flex-row flex items-center mr-3 mb-2 px-3 py-1 border-1 rounded-full border-sky-300 select-none bg-sky-50'>
-          <span className='mr-1 text-sky-300'>{name}</span>
-        </div>
-    )
-  }
-  return (
-      <div className='flex-row flex items-center px-3 py-1 border-1 rounded-full border-fuchsia-300 select-none bg-fuchsia-50'>
-        <span className='mr-1 text-fuchsia-300'>{name}</span>
-      </div>
-  )
 }
